Add re-indexing when a content channel item changes
Vectors can be recalculated for a particular content channel item
when the item changes. This uses Rock's event bus to listen for
changes to content channel items that are indexed. This change
required moving the index configuration out of the Rock job so that
it could be accessed ad-hoc when content channel items change, so
there is a new block for editing the configuration from the
Installed Plugins page.